Generalized ItemWindow Filters
+ Added dropdown for filter selection (draws from Resources.structure_arrays)
+ Added ability for some filters to act as simple booleans. (i.e. checking the Exclusive filter filters out all but Exclusive events)
+ Fixed the exclusive checkbox in the EventItemEditor.  It previously unchecked itself immediately after being checked.

TODO:  Add a more thorough dropdown that allows for multiple filters-- perhaps reuse the ItemWindow for this purpose?  The filter in the sub-item-window could be made to be inactive for this purpose.  Alternatively, the filters and the item window could be made separate to add modularity.  Explore what kind of exclusivity/inclusivity this allows, but allow explore adding logical operations to these filters (and / or / xor?).